# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Car Rental API
  version: v1
  description: Contract summary. Runtime Swagger is generated by springdoc at /v3/api-docs.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/auth/register:
    post:
      security: []
      summary: Register a customer
      responses:
        "201": { description: JWT issued }
  /api/v1/auth/login:
    post:
      security: []
      summary: Login
      responses:
        "200": { description: JWT issued }
  /api/v1/vehicles/search:
    get:
      summary: Find cars available during a time range (admins also see reserved cars)
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [SEDAN, SUV, VAN] }
        - in: query
          name: latitude
          schema: { type: number, minimum: -90, maximum: 90 }
        - in: query
          name: longitude
          schema: { type: number, minimum: -180, maximum: 180 }
        - in: query
          name: startDateTime
          description: Start of the requested rental period; must be supplied with endDateTime
          schema: { type: string, format: date-time }
        - in: query
          name: endDateTime
          description: Exclusive end of the requested rental period; must be after startDateTime
          schema: { type: string, format: date-time }
      responses:
        "200": { description: Available vehicles }
        "400": { description: Invalid/incomplete position or time range }
  /api/v1/vehicles/{vehicleId}/state:
    get:
      summary: Get vehicle state
      parameters:
        - $ref: "#/components/parameters/vehicleId"
      responses:
        "200": { description: Vehicle state }
  /api/v1/vehicles/{vehicleId}/position:
    patch:
      summary: Vehicle device updates GPS position
      parameters:
        - $ref: "#/components/parameters/vehicleId"
      responses:
        "200": { description: Updated vehicle }
  /api/v1/vehicles/{vehicleId}/fuel:
    patch:
      summary: Vehicle device reports fuel; backend calculates remaining driving range
      parameters:
        - $ref: "#/components/parameters/vehicleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fuelLiters]
              properties:
                fuelLiters: { type: number, minimum: 0, example: 21.0 }
                odometerKm: { type: number, minimum: 0, example: 22120.0 }
      responses:
        "200": { description: Updated vehicle }
  /api/v1/admin/vehicles:
    post:
      summary: Admin adds a car; nearest active branch is derived from its position
      responses:
        "201": { description: Car created }
  /api/v1/admin/vehicles/{vehicleId}:
    delete:
      summary: Admin retires a car
      parameters:
        - $ref: "#/components/parameters/vehicleId"
      responses:
        "204": { description: Car retired }
  /api/v1/reservations:
    post:
      summary: Reserve a car type for a date/time and number of days
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, maxLength: 100 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [carType, startDateTime, numberOfDays]
              properties:
                carType: { type: string, enum: [SEDAN, SUV, VAN] }
                startDateTime: { type: string, format: date-time }
                numberOfDays: { type: integer, minimum: 1, maximum: 365 }
      responses:
        "201": { description: Scheduled reservation created }
        "409": { description: Requested type unavailable for the period }
  /api/v1/reservations/{reservationId}/cancel:
    post:
      summary: Cancel a hold
      parameters:
        - in: path
          name: reservationId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Reservation cancelled }
  /api/v1/rides/start:
    post:
      summary: Start ride
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reservationId]
              properties:
                reservationId: { type: string, format: uuid }
      responses:
        "201": { description: Ride started }
  /api/v1/rides/{rideId}:
    get:
      summary: Get ride state
      parameters:
        - $ref: "#/components/parameters/rideId"
      responses:
        "200": { description: Ride state }
  /api/v1/rides/{rideId}/finish:
    post:
      summary: Finish ride at a selected return branch and record payment
      parameters:
        - $ref: "#/components/parameters/rideId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [returnLocationId]
              properties:
                returnLocationId: { type: string, format: uuid }
                paymentReference: { type: string, maxLength: 100 }
      responses:
        "200": { description: Finished ride with pickup/return branches and payment record }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    vehicleId:
      in: path
      name: vehicleId
      required: true
      schema: { type: string, format: uuid }
    rideId:
      in: path
      name: rideId
      required: true
      schema: { type: string, format: uuid }